Refactor - Made shared class file names pascal case for consistency and fixed requires accordingly.
