+++ shared/HttpErrors.js
